Extended unicode character support (#694)
## What is the goal of this PR?

We've added support for characters in Studio's text editor larger than 16 bits.

## What are the changes implemented in this PR?

Some definitions for clarity:
- 'Char': A Kotlin Char. Fixed to 16 bits, and so are unable to fully encapsulate Unicode codepoints larger than 65,535.
- 'Glyph': A logical character. Each of the following are individual single glyphs, comma separated: `[ a, A, &, 😍, 𐩂, 𑓇, 誺]` 

We've introduced a `GlyphLine` class, which wraps an `AnnotatedString`. `GlyphLine` exposes a few methods used to translate the information Compose Multiplatform (Studio's GUI framework) gives to us based on the assumption that each `Glyph` is one Kotlin `Char`:
1) `subSequenceSafely`
2) `glyphToCharOffset`
3) `charToGlyphOffset`
4) `indexOf`

We've also made changes to the `TextRendering` class. Instead of invalidating the `TextLayoutResult`s of all lines upon an update, we only invalidate the lines that are changed, that being either an insertion or a deletion. This has allowed us to clean up the internal class, removing details from the previous versioning system.

Closes https://github.com/vaticle/typedb-studio/issues/593