Remove export schema button from database manager, unused kotlin loggers (#718)
## What is the goal of this PR?

We've removed the ability to export schema from the database manager, as
this isn't the appropriate place for the functionality (the database
manager is solely for the management of databases). We plan to
reintroduce this functionality into the menu bar when it is implemented.

We've also removed many unusued Kotlin loggers.

## What are the changes implemented in this PR?

We've removed the export schema button from the database manager. 

We've removed unused Kotlin loggers.