Improve password expiry notification (#722)
## What is the goal of this PR?

We've updated the version of typedb-client-java being depended upon,
adding support for `passwordExpirySeconds`, a new user API that replaces
`passwordExpiryDays`.

We've also improved the password expiry notification, improving its
clarity.

## What are the changes implemented in this PR?

We've replaced the implementation that made use of `passwordExpiryDays`
with `passwordExpirySeconds`. We now represent the amount of time until
the password expires as a duration, as this is a more canonical
representation of the time until an event.

We've also improved the clarity of the notification, making it more
accurate and eliminating an edge case where we would say that the
password will expire 'within 0 hours'.