Minimise queries made to refresh state after a schema write (#771)
## What is the goal of this PR?

We no longer reload the entire schema and all its connections on each
schema write. Instead, we now reload the type hierarchy (for the Type
Browser), and also, while there is a Type Editor open, it reloads all
its data on a schema write.

This change drastically reduces the time taken to complete schema write
operations.

## What are the changes implemented in this PR?

There was a bug (#723) where schema writes would freeze up Studio for a
long time - specifically, it wouldn't be possible to commit the
transaction or perform further queries. This was caused by overly
aggressive refresh logic; we recursively loaded every edge for every
type in the whole schema, using Studio's global schema write
transaction. This rendered it unusable for anything until that refresh
was done.

Now, we only reload the type hierarchy (necessary for Type Browser).

Also, the Schema service now broadcasts a "schema write" event whenever
a schema write occurs. When opening a Type Editor, we subscribe to
schema write events. On receiving an event, we reload the Type Editor.
On closing the Type Editor, we unsubscribe from these events.

- closes #723