Fix unpacking driver runtime files into working directory (#789)
## What is the goal of this PR?

Due to a bug in the Java driver, we unpacked the driver dynamic library
into the current working directory of Studio, rather than into a
temporary directory. This was at best confusing, and at worst caused an
unrecognized error in cases where the Studio was run from a read-only
directory (such as `/Applications/` on mac OS.

We update to the driver version with the fix.