In text editor, fix behaviour of Home and End keys (#836)
## Usage and product changes

The behaviour of Home and End keys are now in line with most other text
editors, moving to the start/end of the line (rather than the file)
unless Ctrl (Cmd on MacOS) is additionally pressed.

## Implementation

We had some confusing terminology in the `Command` class - `MOVE_HOME`
and `MOVE_END` intuitively sounded like they should correspond to a
keyboard's Home and End keys, but this resulted in incorrect behaviour.
We've renamed them to `MOVE_CONTENT_START` and `MOVE_CONTENT_END`, and
these are now correctly only triggered by Ctrl+Home and Ctrl+End
respectively. Home and End take you to the start/end of a line. If Shift
is additionally pressed in any case, trigger selection.