Add sentry error reporting for all fatal errors (#815)
## Usage and product changes

We add anonymous Sentry error reporting for all fatal errors in TypeDB
Studio. The Studio version and a hash of the mac address (if available)
are reported along with the error.

Users can opt-out of the diagnostics reporting with a new checkbox in
the `Preferences -> System` checkbox. Note that a Studio restart is
required for the toggle to take effect.

## Implementation

1) We depend on Sentry via maven
2) Configure Diagnostics reporting to Sentry's TypeDB Studio endpoint
3) Generate and inject the version number into a source file, which is
read and included into TypeDB studio