Deploy APT bundles (#839)
## Usage and product changes

TypeDB Studio now deploys APT bundles to `repo.typedb.com` on each
release. Both x86_64 and arm64 architectures are supported.

## Implementation

Given that an APT bundle is just a DEB bundle, we simply take our
existing DEB bundle (created by the `:assemble-platform` target when
built on Debian / Linux) and deploy it directly to APT via a
`deploy_apt` target.

`assemble-platform` can only compile a bundle for the processor
architecture and OS of the machine it is run on. Therefore, for now, we
delegate to CircleCI in order to support `linux-arm64`.