Allow connection through Connection URIs (#843)
## Usage and product changes

We introduce the ability to connect to TypeDB through a "Connection
URI".

It woudl contain the requisite information to connect to TypeDB, that
is:
- Server type (core vs cloud)
- For core:
  - A single address
- For cloud:
  - A username/password pair
  - One or more addresses, which may or may not be translated
  - An optional query parameter indicating if TLS should be enabled

## Implementation

We update the `ServerDialog` to have two modes: Connection URI and
Advanced config. Advanced config is the same as previously, while
Connection URI is a multiline input field. The contents of the
Connection URI field while be synchronised with those of the Advanced
Config fields.