Separate address parsing logic for Core and Cloud connection URIs (#849)
## Usage and product changes

We fix an issue where attempting to parse `typedb-core://` connection
URIs would cause crashes

## Motivation

This issue made this feature unusable for TypeDB Core connection URIs

## Implementation

We separate the logic for parsing addresses for TypeDB Cloud - which
expects a username and password - from the address parsing logic for
TypeDB Core - which doesn't. Instead of attempting to extract data it
doesn't need from the `authority` component of the URI, which resulted
in an empty list of addresses, we simply take the entire `authority`
component as the address.