Allow URI schemes in addresses (#851)
## Usage and product changes

We allow the use of URI schemes (for example, `https://` or `http://`)
in connection addresses

## Motivation

Currently, our driver assumes `http://` if no scheme is supplied -
sometimes, another scheme (such as `https://`) may be required. This is
not possible through studio.

## Implementation

Two points are required to make this functional:
- Correct the `addressFormatIsValid` check to not view the `:` in a
scheme as being a port delimiter. We achieve this through a regex with a
negative lookahead, so that rather than splitting on `:` it splits on
`:` not followed by `//`
- Rewrite the ConnectionURI parser to support schemes in addresses. We
can no longer use the built-in `URI` class to parse it, as that fails
when the addresses have a scheme. Similarly to address format checks, we
make use of a regex with a negative lookbehind so that when we split on
`/` to get the path, we only split on `/` not preceded by `/` or `:`