Update TypeDB driver dependency and constructor (#857)
## Usage and product changes
Update TypeDB driver dependency and use the unified driver constructor
to support the latest API.

## Implementation
When Cloud mode is used, we just pick the first selected address and
send the request (for address translation, it will probably fail the
request with a connection error). It's the easiest thing we can do
considering that the UI is not locked in 3.x even if the feature is not
available.