Add schema visualization tool (#868)
## Release notes: product changes

The new schema visualization tool allows you to view your schema in
graph format.

## Motivation

Currently, after connecting to your TypeDB server, you get stuck - you
don't know what the schema is and it isn't being presented to you.

The richest representation is as a graph on a standalone page, which
this PR adds. We'd also love to show it in the query tool, and will add
this at a later date.

## Implementation

### Overview

We introduce `SchemaState`, a global service responsible for providing a
snapshot of the schema of the connected database.

### How we load schema

The schema is currently retrieved through four TypeQL queries whose
outputs are aggregated into a single visualization.

1. `match { $t sub! $supertype; } or {$t sub $supertype; $t is
$supertype; };`
2. `match { $t owns $attr; not { $t sub! $sown; $sown owns $attr; }; };`
3. `match { $t relates $related; not { $t sub! $srel; $srel relates
$related; }; };`
4. `match { $t plays $played; not { $t sub! $splay; $splay plays
$played; }; };`

We hope to reduce this to one query once all TypeDB versions supported
by Studio include a fix to the following Core issue:
- https://github.com/typedb/typedb/issues/7475 .

### When we load schema

We refresh schema state when:
1. The user connects to a new database;
2. The user hits the "Refresh" button on the Schema page.
3. _(not implemented yet)_ The user hits the "Refresh" button in the
Schema tool window on the Query page.

We do NOT refresh schema state when a schema transaction is committed;
we leave this up to the user. This is because refreshing the schema
state destroys the graph visualization state. The user may have done
dragging and dropping to get it to a shape they like.

Besides, even if we did refresh in that scenario, this would represent
an incomplete solution, as TypeDB does not yet provide any mechanism for
notifying clients of schema changes (performed by other clients)

### How we persist the visualization across page navigations

The visualization on the Schema page persists across in-app page
navigation. We don't get this for free - a Sigma WebGL context is
destroyed when its canvas DOM element is, and this happens when you
leave a page. So, we save the graph state on destroying the canvas, and
restore it when it comes back.