Fix autocomplete suggesting erratically & add more suggestions (#866)
## Release notes: product changes
We fix a bug where the suggestions of variables & labels were
unpredictable, and implement more suggestions in an extensible, but
ad-hoc way.
We enable suggestions based on a datastructure representing the schema,
allowing labels to be suggested based on the schema. The schema is
loaded both from the database and from the editor state.

## Motivation
Better autocomplete.

## Implementation
We walk up the tree till we find a node we want to do a suggestion at.
We're provided with a "prefix". This is the union of the direct children
of every node on the path from this suggesting node to the node being
parsed*. The nodes on the path are excluded.
E.g. If we reach `ClauseMatch` where `ClauseMatch { MATCH Patterns };
Pattern { (Pattern SEMICOLON)+ }` on our climb,
`match $x isa person; $y `, prefix would be `[MATCH, Pattern, SEMICOLON,
VAR]`.

* (Any children to the right of th e parse point are ignored, because we
could be editing in-between nodes)

Ideally, we'd be able to partially generate suggestions from the
grammar. Lezer provides a grammar for lezer grammar files.

### Manually testing the schema-based suggestions
```
# Run the following query.
match $default-owner owns $default-owned; limit 1;
match $default-relation relates $default-related; limit 1;
match $default-player plays $default-played; limit 1;
match

{ $owner owns $owned; $relation is $default-relation; $related is $default-related; $player is $default-player; $played is $default-played; } or 
{ $owner is $default-owner; $owned is $default-owned; $relation relates $related; $player is $default-player; $played is $default-played; } or
{ $owner is $default-owner; $owned is $default-owned; $relation is $default-relation; $related is $default-related; $player plays $played; };
```
And this in the JS console:
`_updateSchemaFromDB(_lastQueryAnswers, _lastQueryAnswers,
_lastQueryAnswers)`