Make TAB insert a tab character in the query pane (#883)
## Release notes: product changes

The Tab key now inserts a tab character in the query pane.

## Motivation

This should drastically improve the editing experience for a majority of
users.

## Implementation

Import the `indentWithTab` standard CM6 command.

## Additional Information

This change is a violation of W3C standards that may degrade
accessibility for a few users. We are balancing the cost of implementing
accessibility properly against overall user experience with this change.