Fix a bug where schema vertices with no edges were not rendered in graph output (#885)
## Release notes: product changes

We fixed a bug where schema vertex "islands" (with no edges) were not
rendered in graph output.

## Motivation

It was confusing.

## Implementation

We add `put_kind` to the graph visualiser framework's
`ILogicalGraphConverter` allowing it to handle query constraints tagged
`kind`. When Studio's graph output in the Query pane sees a `kind` query
constraint, we add a matching schema vertex to the graph for the schema
type that it describes (if not already present).

## Additional Information

<img width="1393" height="852" alt="image"
src="https://github.com/user-attachments/assets/d6e9b564-768f-4aa0-99b0-77100723da6f"
/>