Add validation error message for HTTP server address in Safari (#892)
## Release notes: product changes

Add a validation error message in Safari when attempting to connect to
an HTTP endpoint rather than HTTPS.

## Motivation

Safari forbids mixed content, viz. in our case HTTP requests from an
HTTPS page as a security concern. This means that a user of
https://studio.typedb.com/ will not be able to connect to the HTTP
endpoint (usually http://localhost:8000/) of a locally running TypeDB CE
server.

## Implementation

Added a `safariMixedContentValidator` to the address field in connection
dialogue.