Improve connection validation (#891)
## Release notes: product changes

When the port is not written, we issue a warning, instead of an error
(indicate to the user the default ports will be used).

We also update the Safari rejection wrarning that was too strict, since
Chrome on Mac may also include "Safari" in the user agent (?)

```
Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36
```

## Motivation

Observed a user not knowing that they must write the port!

## Implementation

AI extended the existing validation with one new one.